+++ pkg/util/aws/acm.go
@@ -1,4 +1,4 @@
-package awsutil
+package aws
 
 import (
 	"github.com/aws/aws-sdk-go/aws/session"
