+++ pkg/util/aws/ec2.go
@@ -1,4 +1,4 @@
-package awsutil
+package aws
 
 import (
 	"fmt"
