+++ pkg/util/aws/iam.go
@@ -1,4 +1,4 @@
-package awsutil
+package aws
 
 import (
 	"strings"
