+++ pkg/util/aws/route53_test.go
@@ -1,4 +1,4 @@
-package awsutil
+package aws
 
 import (
 	//"fmt"
